iq-trading-bot/requirements.txt
iqoptionapi
stable-baselines3
torch
mplfinance
pandas-ta
python-dotenv
pyyaml
gymnasium
sqlalchemy
